Refactor: combine all 'belong' relation to one table
Because, if we didn't, we'd need a lot of relational tables to represent entity relationships.

For example, Tag have two 'belong' relation, the first one is Category, another is Subject.
Therefore, usually we need two relation table to represent this relationship.
However, when it comes to maintenance, this is a bit overwhelming.